requirements.txt
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.2
pyyaml>=6.0.1
jinja2>=3.1.2
python-dateutil>=2.8.2
